Moving IastInstrumenter to a new module (iast-instrumenter) inside dd-java-agent/instrumentation
