+++ dd-java-agent/agent-tooling/agent-tooling.gradle
@@ -38,9 +38,6 @@ dependencies {
   testImplementation group: 'com.google.guava', name: 'guava-testlib', version: '20.0'
 
   jmhImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-web', version: '2.3.5.RELEASE'
-
-  annotationProcessor deps.autoserviceProcessor
-  compileOnly deps.autoserviceAnnotation
 }
 
 jmh {
