+++ dd-java-agent/instrumentation/iast-instrumenter/iast-instrumenter.gradle
@@ -1,0 +1,18 @@
+muzzle {
+  pass {
+    coreJdk()
+  }
+}
+
+apply from: "${rootDir}/gradle/java.gradle"
+apply plugin: 'org.unbroken-dome.test-sets'
+apply plugin: 'call-site-instrumentation'
+
+testSets {
+  latestDepTest {
+    dirName = 'test'
+  }
+}
+
+dependencies {
+}
