+++ dd-java-agent/instrumentation/iast-instrumenter/src/test/groovy/IastInstrumenterTest.groovy
@@ -1,6 +1,5 @@
-package datadog.trace.agent.tooling
-
-
+import datadog.trace.agent.tooling.Instrumenter
+import datadog.trace.instrumentation.iastinstrumenter.IastInstrumenter
 import datadog.trace.test.util.DDSpecification
 import net.bytebuddy.description.type.TypeDescription
