+++ dd-java-agent/instrumentation/servlet/request-2/request-2.gradle
@@ -41,6 +41,8 @@ dependencies {
   testImplementation group: 'org.eclipse.jetty', name: 'jetty-server', version: '7.0.0.v20091005'
   testImplementation group: 'org.eclipse.jetty', name: 'jetty-servlet', version: '7.0.0.v20091005'
 
+  testImplementation project(':dd-java-agent:instrumentation:iast-instrumenter')
+
   latestDepTestImplementation group: 'org.eclipse.jetty', name: 'jetty-server', version: '7.+'
   latestDepTestImplementation group: 'org.eclipse.jetty', name: 'jetty-servlet', version: '7.+'
 }
