+++ settings.gradle
@@ -182,6 +182,7 @@ include ':dd-java-agent:instrumentation:hibernate:core-4.0'
 include ':dd-java-agent:instrumentation:hibernate:core-4.3'
 include ':dd-java-agent:instrumentation:http-url-connection'
 include ':dd-java-agent:instrumentation:hystrix-1.4'
+include ':dd-java-agent:instrumentation:iast-instrumenter'
 include ':dd-java-agent:instrumentation:ignite-2.0'
 include ':dd-java-agent:instrumentation:java-security'
 include ':dd-java-agent:instrumentation:jakarta-rs-annotations-3'
